Add coverpkg to track coverage
